define Reation and RoomReaction in reaction.ts, consistent with Message.ts and Messages.ts
