+++ src/RoomReactions.ts
@@ -2,38 +2,9 @@ import * as Ably from 'ably';
 
 import { RoomReactionEvents } from './events.js';
 import { Logger } from './logger.js';
+import { DefaultReaction, Reaction } from './Reaction.js';
 import { SubscriptionManager } from './SubscriptionManager.js';
 import EventEmitter from './utils/EventEmitter.js';
-
-/**
- * Represents a room-level reaction.
- */
-export interface Reaction {
-  /**
-   * The type of the reaction, for example "like" or "love".
-   */
-  readonly type: string;
-
-  /**
-   * metadata of the reaction, if any was set
-   */
-  readonly metadata?: any;
-
-  /**
-   * The timestamp at which the reaction was sent.
-   */
-  readonly createdAt: Date;
-
-  /**
-   * The clientId of the user who sent the reaction.
-   */
-  readonly clientId: string;
-
-  /**
-   * Whether the reaction was sent by the current user.
-   */
-  readonly isSelf: boolean;
-}
 
 /**
  * The listener function type for room-level reactions.
@@ -196,18 +167,5 @@ export class DefaultRoomReactions extends EventEmitter<RoomReactionEventsMap> im
       inbound.clientId === clientId,
       inbound.data.metadata,
     );
-  }
-}
-
-class DefaultReaction implements Reaction {
-  constructor(
-    public readonly type: string,
-    public readonly clientId: string,
-    public readonly createdAt: Date,
-    public readonly isSelf: boolean,
-    public readonly metadata: any,
-  ) {
-    // The object is frozen after constructing to enforce readonly at runtime too
-    Object.freeze(this);
   }
 }
